requirements.txt
ttkbootstrap==1.10.1
Pillow==8.3.2
pycountry==20.7.3
tkcalendar==1.6.1